+++ packages/server/database/rethinkDriver.ts
@@ -119,10 +119,6 @@ export type RethinkSchema = {
     type: PushInvitation
     index: 'userId'
   }
-  QueryMap: {
-    type: any
-    index: ''
-  }
   MeetingTemplate: {
     type: MeetingTemplate
     index: 'teamId' | 'orgId'
@@ -155,10 +151,6 @@ export type RethinkSchema = {
       | 'discussionId'
       | 'userId'
       | 'integrationHash'
-  }
-  TaskHistory: {
-    type: any
-    index: 'taskIdUpdatedAt' | 'teamMemberId'
   }
   TeamInvitation: {
     type: TeamInvitation