+++ packages/server/graphql/mutations/updateTeamName.ts
@@ -1,7 +1,6 @@
 import {GraphQLNonNull} from 'graphql'
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
 import teamNameValidation from 'parabol-client/validation/teamNameValidation'
-import getTeamsByIds from '../../postgres/queries/getTeamsByIds'
 import updateTeamByTeamId from '../../postgres/queries/updateTeamByTeamId'
 import {analytics} from '../../utils/analytics/analytics'
 import {getUserId, isTeamMember} from '../../utils/authorization'
@@ -36,11 +35,10 @@ export default {
     }
 
     // VALIDATION
-    const [teams, viewer] = await Promise.all([
-      getTeamsByIds([teamId]),
+    const [team, viewer] = await Promise.all([
+      dataLoader.get('teams').loadNonNull(teamId),
       dataLoader.get('users').loadNonNull(viewerId)
     ])
-    const team = teams[0]!
     const oldName = team.name
     const newName = updatedTeam.name
     const orgTeams = await dataLoader.get('teamsByOrgIds').load(team.orgId)
@@ -59,6 +57,7 @@ export default {
       updatedAt: now
     }
     await updateTeamByTeamId(dbUpdate, teamId)
+    dataLoader.clearAll('teams')
     analytics.teamNameChanged(viewer, teamId, oldName, newName, oldName.endsWith('’s Team'))
 
     const data = {teamId}