+++ packages/server/graphql/public/mutations/acceptRequestToJoinDomain.ts
@@ -1,12 +1,11 @@
+import {sql} from 'kysely'
 import DomainJoinRequestId from 'parabol-client/shared/gqlIds/DomainJoinRequestId'
 import {InvoiceItemType, SubscriptionChannel} from 'parabol-client/types/constEnums'
 import toTeamMemberId from 'parabol-client/utils/relay/toTeamMemberId'
+import TeamMemberId from '../../../../client/shared/gqlIds/TeamMemberId'
 import adjustUserCount from '../../../billing/helpers/adjustUserCount'
-import getRethink from '../../../database/rethinkDriver'
 import getKysely from '../../../postgres/getKysely'
-import getTeamsByIds from '../../../postgres/queries/getTeamsByIds'
 import {getUserById} from '../../../postgres/queries/getUsersByIds'
-import addTeamIdToTMS from '../../../safeMutations/addTeamIdToTMS'
 import insertNewTeamMember from '../../../safeMutations/insertNewTeamMember'
 import {Logger} from '../../../utils/Logger'
 import RedisLock from '../../../utils/RedisLock'
@@ -14,6 +13,7 @@ import {getUserId} from '../../../utils/authorization'
 import publish from '../../../utils/publish'
 import setUserTierForUserIds from '../../../utils/setUserTierForUserIds'
 import standardError from '../../../utils/standardError'
+import isValid from '../../isValid'
 import {MutationResolvers} from '../resolverTypes'
 
 // TODO (EXPERIMENT: prompt-to-join-org): some parts are borrowed from acceptTeamInvitation, create generic functions
@@ -27,7 +27,6 @@ const acceptRequestToJoinDomain: MutationResolvers['acceptRequestToJoinDomain'] 
   const viewerId = getUserId(authToken)
   const now = new Date()
   const pg = getKysely()
-  const r = await getRethink()
 
   // Fetch the request that is not expired
   const request = await pg
@@ -44,21 +43,20 @@ const acceptRequestToJoinDomain: MutationResolvers['acceptRequestToJoinDomain'] 
   const {createdBy, domain} = request
 
   // Viewer must be a lead of the provided teamIds
-  const validTeamMembers = await r
-    .table('TeamMember')
-    .getAll(r.args(teamIds), {index: 'teamId'})
-    .filter({isLead: true, userId: viewerId})
-    .run()
+  const viewerTeamMembers = await dataLoader.get('teamMembersByUserId').load(viewerId)
+  const validTeamMembers = viewerTeamMembers.filter(
+    ({isLead, teamId}) => isLead && teamIds.includes(teamId)
+  )
 
   if (!validTeamMembers.length) {
     return standardError(new Error('Not a team leader'))
   }
 
   // Provided request domain should match team's organizations activeDomain
-  const leadTeams = await getTeamsByIds(validTeamMembers.map((teamMember) => teamMember.teamId))
-  const teamOrgs = await Promise.all(
-    leadTeams.map((t) => dataLoader.get('organizations').loadNonNull(t.orgId))
-  )
+  const leadTeams = (await dataLoader.get('teams').loadMany(teamIds)).filter(isValid)
+  const orgIds = leadTeams.map((team) => team.orgId)
+  const teamOrgs = (await dataLoader.get('organizations').loadMany(orgIds)).filter(isValid)
+
   const validOrgIds = teamOrgs.filter((org) => org.activeDomain === domain).map(({id}) => id)
 
   if (!validOrgIds.length) {
@@ -86,14 +84,32 @@ const acceptRequestToJoinDomain: MutationResolvers['acceptRequestToJoinDomain'] 
     return standardError(new Error('User not found'))
   }
 
-  const {id: userId} = user
+  const {id: userId, picture, preferredName, email} = user
 
   for (const validTeam of validTeams) {
     const {id: teamId, orgId} = validTeam
     const [organizationUser] = await Promise.all([
       dataLoader.get('organizationUsersByUserIdOrgId').load({orgId, userId}),
-      insertNewTeamMember(user, teamId),
-      addTeamIdToTMS(userId, teamId)
+      pg
+        .with('UserUpdate', (qc) =>
+          qc
+            .updateTable('User')
+            .set({tms: sql`arr_append_uniq("tms", ${teamId})`})
+            .where('id', '=', userId)
+        )
+        .insertInto('TeamMember')
+        .values({
+          id: TeamMemberId.join(teamId, userId),
+          teamId,
+          userId,
+          picture,
+          preferredName,
+          email,
+          openDrawer: 'manageTeam'
+        })
+        .onConflict((oc) => oc.column('id').doUpdateSet({isNotRemoved: true}))
+        .execute(),
+      insertNewTeamMember(user, teamId, dataLoader)
     ])
 
     if (!organizationUser) {
@@ -105,7 +121,7 @@ const acceptRequestToJoinDomain: MutationResolvers['acceptRequestToJoinDomain'] 
       await setUserTierForUserIds([userId])
     }
   }
-
+  dataLoader.clearAll(['users', 'teamMembers'])
   await redisLock.unlock()
 
   // Send the new team member a welcome & a new token